# storage/api-docs/api-docs.json
{
    "swagger": "2.0",
    "info": {
        "title": "API Documentation",
        "description": "Server endpoint for Casper project. Each of the resources below has multiple endpoints which allows end users and developers to access the Gunther API. This server allows you can use the api key  \"special-key\" to test the authorization filters",
        "termsOfService": "http://localhost",
        "contact": {
            "email": ""
        },
        "license": {
            "name": "",
            "url": ""
        },
        "version": "0.0.1"
    },
    "host": "localhost:9000",
    "basePath": "/v1/api",
    "schemes": [
        "http"
    ],
    "paths": {
...
        "/products": {
            "get": {
                "tags": [
                    "Product"
                ],
                "summary": "Get all products",
                "description": "Returns all products.",
                "operationId": "get.products",
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "Retrieve all products.",
                        "schema": {
                            "$ref": "#/definitions/Product"
                        }
                    },
                    "404": {
                        "description": "Error fetching products.",
                        "schema": {
                            "$ref": "#/definitions/ErrorModel"
                        }
                    }
                }
            }
        },
...
        "/products/add": {
            "post": {
                "tags": [
                    "Product"
                ],
                "summary": "Create a product",
                "description": "Creates a new product",
                "operationId": "create.product",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "name",
                        "in": "formData",
                        "description": "Name of product to be created",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "description",
                        "in": "formData",
                        "description": "Product description",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "price",
                        "in": "formData",
                        "description": "Price of product to be created",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "imageurl",
                        "in": "formData",
                        "description": "Image url of product to be created",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "file",
                        "in": "formData",
                        "description": "Product image file to upload",
                        "required": false,
                        "type": "file"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of products",
                        "schema": {
                            "$ref": "#/definitions/ErrorModel"
                        }
                    },
                    "404": {
                        "description": "Error creating product",
                        "schema": {
                            "$ref": "#/definitions/ErrorModel"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "Product": {
            "required": [
                "name",
                "description",
                "price",
                "imageurl",
                "file_id"
            ],
            "properties": {
                "id": {
                    "description": "Unique product identifier",
                    "type": "string"
                },
                "name": {
                    "description": "Product name",
                    "type": "string"
                },
                "description": {
                    "description": "Product description",
                    "type": "string"
                },
                "price": {
                    "description": "Product price",
                    "type": "string"
                },
                "imageurl": {
                    "description": "Product image url",
                    "type": "string"
                },
                "file_id": {
                    "description": "Product file_id",
                    "type": "string"
                }
            },
            "type": "object",
            "xml": {
                "name": "Product"
            },
            "externalDocs": {
                "description": "find more info here",
                "url": "https://swagger.io/about"
            }
        }
    },
...
}